+++ packages/codemirror-block-editor/src/demo/write-notifications.ts
@@ -1,0 +1,10 @@
+export function writeNotificationToHtml(text: string) {
+  const container = document.getElementById("notifications");
+  if (!container) {
+    console.error("Could not find notifications container.");
+    return;
+  }
+  const notificationEl = document.createElement("div");
+  notificationEl.textContent = text;
+  container.prepend(notificationEl);
+}